pwm1 working

+++ firmware/src/edubrm.h
@@ -1,4 +1,4 @@
 void GetInReport (uint8_t src[], uint32_t length);
 void SetOutReport (uint8_t dst[], uint32_t length);
-void EnablePWM1(uint16_t period, uint16_t duty);
+void EnablePWM1(uint32_t period, uint32_t duty);
 void EnablePWM2(uint16_t period, uint16_t duty);
